Remove query params from launch params.
When @jsphmrrw and I were running through the IMS LTI certification we
ran into an issue relating to how query parameters are included in the
launch parameters.

The example launch url used by the IMS LTI certification suite includes
query params (x and y).

As the Tool consumer you need to generate an OAuth signature using your
credentials and the LTI parameters.

When you hand the user over to the Tool provider, you issue a POST
request, which includes the LTI parameters and OAuth parameters.

Previously the gem would generate a form including the original LTI
parameters, the query parameters and the OAuth parameters, however the
LTI spec does not expect the form to include these query parameters.

This commit updates the `generate_launch_data` method so that query
parameters are stripped out of the hash it returns.

See this thread for more details:

https://community.canvaslms.com/thread/19405-verifying-signature-in-lti-launch-when-there-are-query-parameters